+++ packages/liquidation-sdk-viem/src/index.ts
@@ -4,6 +4,8 @@ export * from "./api/index.js";
 export * from "./addresses.js";
 export * from "./flashbots.js";
 export * from "./LiquidationEncoder.js";
+export * from "./preLiquidation/index.js";
 export * from "./swap/index.js";
 export * from "./tokens/index.js";
 export * from "./abis.js";
+export * from "./positions/index.js";
