+++ packages/liquidation-sdk-viem/src/positions/getters.ts
@@ -1,10 +1,12 @@
 import type { ChainId, MarketId } from "@morpho-org/blue-sdk";
 import { fetchAccrualPosition } from "@morpho-org/blue-sdk-viem";
+import {
+  PreLiquidationPosition,
+  getPreLiquidablePositions,
+} from "@morpho-org/liquidation-sdk-viem";
 import { Time } from "@morpho-org/morpho-ts";
 import type { Account, Chain, Client, Transport } from "viem";
 import { apiSdk } from "../api";
-import { getPreLiquidablePositions } from "../preLiquidation/positionGetters";
-import { PreLiquidationPosition } from "../preLiquidation/types";
 
 export async function getPositions(
   client: Client<Transport, Chain, Account>,
